# clone smallest bitmap, adjust naive/workshy and thresholds in `FastAggregation.and`