This is based on observations made in Apache Pinot and separately in #608 
* for small numbers of bitmaps, naive_and can be faster than the more complex workShyAnd method
* putting the smallest bitmap first in the aggregation can accelerate intersections