@richardstartin 2023-01-03
I don't have time to evaluate this exhaustively, this has been determined to be effective already by Apache Pinot, but ran a quick benchmark on my laptop:

before:
```
Benchmark                         (constructionStrategy)  (count)  (probability)  (seed)   (size)   Mode  Cnt      Score     Error  Units
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10           0.01   99999  1000000  thrpt    5   8469.991 ± 710.023  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10            0.1   99999  1000000  thrpt    5  18620.905 ± 229.084  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10            0.5   99999  1000000  thrpt    5  12077.086 ±  21.410  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100           0.01   99999  1000000  thrpt    5   1444.833 ± 135.439  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100            0.1   99999  1000000  thrpt    5    163.065 ±   3.768  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100            0.5   99999  1000000  thrpt    5    457.596 ±  15.069  ops/s
```

after:
```
Benchmark                         (constructionStrategy)  (count)  (probability)  (seed)   (size)   Mode  Cnt      Score      Error  Units
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10           0.01   99999  1000000  thrpt    5   9597.927 ± 1438.935  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10            0.1   99999  1000000  thrpt    5  18970.302 ±  186.505  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY       10            0.5   99999  1000000  thrpt    5  12065.076 ±   50.169  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100           0.01   99999  1000000  thrpt    5   1459.275 ±   19.018  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100            0.1   99999  1000000  thrpt    5    163.538 ±    2.663  ops/s
FastAggregationRLEStressTest.and         CONSTANT_MEMORY      100            0.5   99999  1000000  thrpt    5    457.225 ±    8.471  ops/s
```