@lemire 2023-01-03
@frensjan Would you be willing to check this out and maybe run a benchmark?