@frensjan 2023-01-04
LGTM!

In my macro benchmarking for a particular query, `FastAggregation.and` used to dominate the benchmark at ~80% in cloning a large bitmap to ~4% in `FastAggregation.and` in total.

In my work I dropped the use of `FastAggregation.and(...)` and incrementally computed the intersection. Something like:

```java
RoaringBitmap postings = null;

for( ... ) {
    RoaringBitmap p = readBitmap(...);
    postings = and(postings, p);
}

private static RoaringBitmap and(RoaringBitmap a, RoaringBitmap b) {
    if (a == null) {
        return b;
    }

    if (a.getLongSizeInBytes() < b.getLongSizeInBytes()) {
        RoaringBitmap result = a.clone();
        result.and(b);
        return result;
    } else {
        RoaringBitmap result = b.clone();
        result.and(a);
        return result;
    }
}
```

This improvement in `FastAggregation` yields slightly faster results in my macro benchmark. The difference seems to stem from the inability to use a cheaper metric to compare the bitmaps on. `getLongSizeInBytes` comes up as a fairly significant method CPU-wise in profiling; `bitmap.highLowContainer.size()` as @richardstartin is using should be way cheaper.

I've also tried a variant with `cardinalityExceeds` but that doesn't make much of a difference.
```java
if (a.cardinalityExceeds(b.getLongCardinality())) {
    ...
`