@richardstartin 2023-01-04
@lemire I'm pretty confident that this isn't going to cause problems so I recommend we merge this