@richardstartin 2023-01-04
> Merging.
> 
> I will issue a release.

The release doesn't seem to have propagated yet, but it looks like the tag was issued 45 minutes ago