@lemire 2023-01-04
> The release doesn't seem to have propagated yet, but it looks like the tag was issued 45 minutes ago

Yes. A new release is upcoming.